Prototype for some Tycho integration in same reactor
This changes does some pare minimal bootstrapping necessary to throw
a bundle recipe project into Tycho and get something back. Previously
it was already possible to generate p2 metadate using Tycho. This change
goes a bit further and actually produces a p2 repository with the
recipe. 

The idea is that instead of relying on a features for collecting things
we just mirror each recipe into a large result repo. This allows for a
future optimization that will rebuild and mirror only recipes that
actually changed.

Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>